Gradio Interface
Wrote a simple webpage interface which allows the user to upload a picture, and decides whether the picture is a fake or a real one using the model that we trained.

The code needs editing for reusability, which I leave for later and/or others.

The trained model which is used here is uploaded to the following link (It cannot be pushed to github):

https://drive.google.com/file/d/1eDAKamvTuD-7_wihSM_0oZYbtMIFxXxR/view?usp=sharing